# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CODE-NFE",
        "description": "Documentação da API gerada com Swagger no Laravel",
        "contact": {
            "email": "brunocostasilva62@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/companys": {
            "get": {
                "tags": [
                    "Rota de empresas que vai utilizar o sistema"
                ],
                "summary": "Lista Todas empresas Cadastrada",
                "operationId": "00db5b4605d1f7d690dd78c799cd0800",
                "parameters": [
                    {
                        "name": "status",
                        "in": "header",
                        "description": "Passar no parametro 1 para ativos e 0 para Inativos'",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista da empresa cadastrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "companey": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "PEPEU 09351223343"
                                                },
                                                "cpf_cnpj": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "41055038000149"
                                                },
                                                "name_fantasy": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "Pepeus Bar LTDA"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "Rua Antonio Mesias"
                                                },
                                                "number_addres": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "10"
                                                },
                                                "district_addres": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "CENTRO"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "Terr Rica"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "PR"
                                                },
                                                "cep": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "87890-000"
                                                },
                                                "inscription_state": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "ISENTO"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "(44) 998212-815"
                                                },
                                                "date_expiration": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "2025-05-10"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rota de empresas que vai utilizar o sistema"
                ],
                "summary": "Cadastro de empresas",
                "operationId": "0ae59365e605aa52e0ed1569ccca8c92",
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf_cnpj"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "PEPEU 09351223343"
                                    },
                                    "cpf_cnpj": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "41055038000149"
                                    },
                                    "name_fantasy": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Pepeus Bar LTDA"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Rua Antonio Mesias"
                                    },
                                    "number_addres": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "10"
                                    },
                                    "district_addres": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "CENTRO"
                                    },
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Terr Rica"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "PR"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "87890-000"
                                    },
                                    "inscription_state": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "ISENTO"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "(44) 998212-815"
                                    },
                                    "name_user": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "bruno Costa"
                                    },
                                    "username": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "bruno2525"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companys-users/{id}": {
            "put": {
                "tags": [
                    "Rota de empresas que vai utilizar o sistema"
                ],
                "summary": "Busca todos os usuarios referente a empresa passado pelo id",
                "operationId": "db370567067f8287df20412e31ecaf32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da empresa que deseja buscar os usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "success": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Bruno Costa"
                                            },
                                            "username": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "bruno1020"
                                            },
                                            "type_user": {
                                                "type": "ENUM",
                                                "format": "text",
                                                "example": "FULL"
                                            },
                                            "status": {
                                                "type": "boolean",
                                                "format": "text",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companys-status/{id}": {
            "put": {
                "tags": [
                    "Rota de empresas que vai utilizar o sistema"
                ],
                "summary": "Se a empresta estiver Ativa e receber essa requicição ela Desativa ou visse versa.",
                "operationId": "7794e97e62aa2ad4cd3e1fcc30817f6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}